app/requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
cryptography==35.0.0
DateTime==4.3
Flask==2.0.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
lxml==4.6.4
MarkupSafe==2.0.1
pdfminer.six==20211012
psycopg2-binary==2.9.3
pycparser==2.21
pycryptodome==3.11.0
PyJWT==2.3.0
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
six==1.16.0
soupsieve==2.3
SQLAlchemy==1.4.29
urllib3==1.26.7
uuid==1.30
Werkzeug==2.0.2
zope.interface==5.4.0
